fix: correctness issues from code audit
- truma_inetbox: QUEUE_LOG_MSG used int8_t for LIN bytes; now uint8_t, re-casts removed
- truma_inetbox: CP Plus binary sensor micros() comparison was not rollover-safe
  (false disconnect possible every ~71 min); use subtraction idiom
- truma_inetbox: get_last_cp_plus_request() returned int64_t for a uint32_t timestamp

Claude

+++ components/truma_inetbox/TrumaiNetBoxApp.h
@@ -39,7 +39,7 @@ class TrumaiNetBoxApp : public LinBusProtocol {
   TrumaiNetBoxAppHeater *get_heater() { return &this->heater_; }
   TrumaiNetBoxAppTimer *get_timer() { return &this->timer_; }
 
-  int64_t get_last_cp_plus_request() { return this->device_registered_.load(std::memory_order_relaxed); }
+  uint32_t get_last_cp_plus_request() { return this->device_registered_.load(std::memory_order_relaxed); }
 
 #ifdef USE_TIME
   void set_time(time::RealTimeClock *time) { time_ = time; }
